Caching Matrix inversion
These two functions cache the inverse of a matrix.The first function is
called 'makeCacheMatrix': This function creates a special "matrix"
object that can cache its inverse.The following function returns the
inverse of the matrix. However, it first checks to see if the inverse
has already been calculated. If so, it gets the inverse of the matrix
from the cache and skips the computation. Otherwise, it calculates the
inverse of the matrix and sets the value of the inverse matrix in the
cache via the setinverse function.